requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-multipart==0.0.6
django-extensions==3.2.3
python-decouple==3.8
matplotlib==3.7.2
seaborn==0.12.2
nltk==3.8.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
